requirements.txt
gsplat>=1.0.0
scikit-learn
tensorboard
pyyaml
easydict
loguru
tqdm
plyfile
pyquaternion
pillow
viser
matplotlib
open3d